Leaking zombie child processes Resolves: bz#106021
Subject: [PATCH] polkitd: fix zombie not reaped when js spawned process timed
 out

The child watch source attached to thread context didn't work due
to the release of it's main loop and context outside. So we attach
the source to the global default main context to make it work and
avoid zombies.